requirements.txt
python-telegram-bot>=20.0
requests
beautifulsoup4
python-dotenv
google-generative-ai